jitfarm_api/requirements.txt
fastapi==0.115.0
fastapi_jwt_auth==0.5.0
SQLAlchemy==1.4.44
passlib==1.7.4
uvicorn==0.19.0
psycopg2
postgres
requests
pymongo
pydantic==1.10.11
pandas
bcrypt
pyjwt[crypto]